+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add `load_queryables` function to pypgstac for loading queryables from a JSON file
+- Add support for specifying collection IDs when loading queryables
+
 ## [v0.9.5]
 
 ### Changed